+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 3.5.0 - 2022-03-10
+### Added
+- support for testing `CKM_XOR_BASE_AND_DATA`
+- support for testing `C_GenerateRandom()`
+- support for testing `C_SeedRandom()`
+
 ## 3.4.1 - 2021-11-30
 ### Fixed
 - `requirements.txt` updated and adjusted to remove unnecessary dependencies